unify how we handle config
