+++ martin/src/pmtiles/mod.rs
@@ -3,8 +3,6 @@ use std::fmt::{Debug, Formatter};
 use std::io;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering::Relaxed;
 
 use async_trait::async_trait;
 use log::{trace, warn};
@@ -16,17 +14,18 @@ use pmtiles::{
     AsyncPmTilesReader, AwsS3Backend, Compression, DirCacheResult, Directory, DirectoryCache,
     HttpBackend, MmapBackend, TileId, TileType,
 };
-use serde::{Deserialize, Serialize};
 use tilejson::TileJSON;
 use url::Url;
 
-use crate::config::UnrecognizedValues;
 use crate::file_config::FileError::{InvalidMetadata, InvalidUrlMetadata, IoError};
-use crate::file_config::{ConfigExtras, FileError, FileResult, SourceConfigExtras};
+use crate::file_config::{FileError, FileResult};
 use crate::source::{TileInfoSource, UrlQuery};
 use crate::utils::cache::get_cached_value;
 use crate::utils::{CacheKey, CacheValue, OptMainCache};
 use crate::{MartinResult, Source, TileData};
+
+mod config;
+pub use config::PmtConfig;
 
 #[derive(Clone, Debug)]
 pub struct PmtCache {
@@ -61,134 +60,6 @@ impl DirectoryCache for PmtCache {
                     CacheValue::PmtDirectory(directory),
                 )
                 .await;
-        }
-    }
-}
-
-#[serde_with::skip_serializing_none]
-#[derive(Debug, Default, Serialize, Deserialize)]
-pub struct PmtConfig {
-    /// Force path style URLs for S3 buckets
-    ///
-    /// A path style URL is a URL that uses the bucket name as part of the path like `example.org/some_bucket` instead of the hostname `some_bucket.example.org`.
-    /// If `None` (the default), this will look at `AWS_S3_FORCE_PATH_STYLE` or default to `false`.
-    #[serde(default, alias = "aws_s3_force_path_style")]
-    pub force_path_style: Option<bool>,
-    /// Skip loading credentials for S3 buckets
-    ///
-    /// Set this to `true` to request anonymously for publicly available buckets.
-    /// If `None` (the default), this will look at `AWS_SKIP_CREDENTIALS` and `AWS_NO_CREDENTIALS` or default to `false`.
-    #[serde(default, alias = "aws_skip_credentials")]
-    pub skip_credentials: Option<bool>,
-    #[serde(flatten)]
-    pub unrecognized: UnrecognizedValues,
-
-    //
-    // The rest are internal state, not serialized
-    //
-    #[serde(skip)]
-    pub client: Option<Client>,
-
-    #[serde(skip)]
-    pub next_cache_id: AtomicUsize,
-
-    #[serde(skip)]
-    pub cache: OptMainCache,
-}
-
-impl PartialEq for PmtConfig {
-    fn eq(&self, other: &Self) -> bool {
-        self.unrecognized == other.unrecognized
-    }
-}
-
-impl Clone for PmtConfig {
-    fn clone(&self) -> Self {
-        // State is not shared between clones, only the serialized config
-        Self {
-            unrecognized: self.unrecognized.clone(),
-            ..Default::default()
-        }
-    }
-}
-
-impl PmtConfig {
-    /// Create a new cache object for a source, giving it a unique internal ID
-    /// and a reference to the global cache.
-    pub fn new_cached_source(&self) -> PmtCache {
-        PmtCache::new(self.next_cache_id.fetch_add(1, Relaxed), self.cache.clone())
-    }
-}
-
-impl ConfigExtras for PmtConfig {
-    fn init_parsing(&mut self, cache: OptMainCache) -> FileResult<()> {
-        assert!(self.client.is_none());
-        assert!(self.cache.is_none());
-
-        self.client = Some(Client::new());
-        self.cache = cache;
-
-        if self.unrecognized.contains_key("dir_cache_size_mb") {
-            warn!(
-                "dir_cache_size_mb is no longer used. Instead, use cache_size_mb param in the root of the config file."
-            );
-        }
-
-        Ok(())
-    }
-
-    fn is_default(&self) -> bool {
-        true
-    }
-
-    fn get_unrecognized(&self) -> &UnrecognizedValues {
-        &self.unrecognized
-    }
-}
-
-impl SourceConfigExtras for PmtConfig {
-    fn parse_urls() -> bool {
-        true
-    }
-
-    async fn new_sources(&self, id: String, path: PathBuf) -> FileResult<TileInfoSource> {
-        Ok(Box::new(
-            PmtFileSource::new(self.new_cached_source(), id, path).await?,
-        ))
-    }
-
-    async fn new_sources_url(&self, id: String, url: Url) -> FileResult<TileInfoSource> {
-        match url.scheme() {
-            "s3" => {
-                let force_path_style = self.force_path_style.unwrap_or_else(|| {
-                    get_env_as_bool("AWS_S3_FORCE_PATH_STYLE").unwrap_or_default()
-                });
-                let skip_credentials = self.skip_credentials.unwrap_or_else(|| {
-                    get_env_as_bool("AWS_SKIP_CREDENTIALS").unwrap_or_else(|| {
-                        // `AWS_NO_CREDENTIALS` was the name in some early documentation of this feature
-                        get_env_as_bool("AWS_NO_CREDENTIALS").unwrap_or_default()
-                    })
-                });
-                Ok(Box::new(
-                    PmtS3Source::new(
-                        self.new_cached_source(),
-                        id,
-                        url,
-                        skip_credentials,
-                        force_path_style,
-                    )
-                    .await?,
-                ))
-            }
-            _ => Ok(Box::new(
-                PmtHttpSource::new(
-                    self.client.clone().unwrap(),
-                    self.new_cached_source(),
-                    id,
-                    url,
-                )
-                .await?,
-            )),
         }
     }
 }
@@ -402,12 +273,4 @@ impl PmtFileSource {
 
         Self::new_int(id, path, reader).await
     }
-}
-
-/// Interpret an environment variable as a [`bool`]
-///
-/// This ignores casing and treats bad utf8 encoding as `false`.
-fn get_env_as_bool(key: &'static str) -> Option<bool> {
-    let val = std::env::var_os(key)?.to_ascii_lowercase();
-    Some(val.to_str().is_some_and(|val| val == "1" || val == "true"))
 }
